- TC#14 - created hooks for carousel, auth, partners and categories components
