+++ frontend/src/features/carousel/BlockCarousel.tsx
@@ -1,7 +1,3 @@
-import { useEffect } from 'react';
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
-import { getCarousel } from '@/features/carousel/CarouselThunk';
-import { loadingCarouselState, photoCarouselState } from '@/features/carousel/CarouselSlice';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
@@ -9,46 +5,11 @@ import { API_URl } from '@/consts';
 import { Loader } from '@/components/Loader/Loader';
 import styles from './Carousel.module.css';
 import './Carousel.css';
+import {useBlockCarousel} from '@/features/carousel/hooks/useBlockCorousel';
 
 export const BlockCarousel = () => {
-  const dispatch = useAppDispatch();
-  const carousel = useAppSelector(photoCarouselState);
-  const loadingCarousel = useAppSelector(loadingCarouselState);
-
-  useEffect(() => {
-    dispatch(getCarousel());
-  }, [dispatch]);
-
-  const settings = {
-    dots: true,
-    infinite: true,
-    speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    autoplay: true,
-    pauseOnFocus: false,
-    dotsClass: 'slick-dots custom-dots',
-    responsive: [
-      {
-        breakpoint: 640,
-        settings: {
-          slidesToShow: 1,
-        },
-      },
-      {
-        breakpoint: 1024,
-        settings: {
-          slidesToShow: 1,
-        },
-      },
-    ],
-  };
-
-  const minimumPhotos = 2;
-  const displayedPhotos =
-    carousel.length >= minimumPhotos
-      ? carousel
-      : [...carousel, ...Array(minimumPhotos - carousel.length).fill({ image: 'placeholder.jpg', _id: 'placeholder' })];
+  
+  const { loadingCarousel, displayedPhotos, settings } = useBlockCarousel();
 
   return (
     <>
