+++ frontend/src/features/partners/AdminPartners.tsx
@@ -1,33 +1,14 @@
 import PartnersCard from '@/features/partners/components/PartnersCard';
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
-import { selectPartners, selectPartnersFetching } from '@/features/partners/partnerSlice';
-import { useEffect, useState } from 'react';
-import { createPartner, fetchPartner } from '@/features/partners/partnerThunks';
 import { Button } from '@/components/ui/button';
 import PartnersForm from '@/features/partners/components/PartnersForm';
-import { Partner } from '@/types/partnerTypes';
 import { Loader } from '@/components/Loader/Loader';
 import { SquaresPlusIcon } from '@heroicons/react/24/outline';
+import {useAdminPartners} from '@/features/partners/hooks/useAdminPartners';
 
 const AdminPartners = () => {
-  const dispatch = useAppDispatch();
-  const partners = useAppSelector(selectPartners);
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const loading = useAppSelector(selectPartnersFetching);
 
-  useEffect(() => {
-    dispatch(fetchPartner());
-  }, [dispatch]);
-
-  const handleSubmit = async (partner: Partner) => {
-    try {
-      await dispatch(createPartner(partner));
-      setIsDialogOpen(false);
-    } catch (error) {
-      console.error('Failed to create partner:', error);
-    }
-  };
-
+  const { partners, isDialogOpen, loading, handleSubmit, setIsDialogOpen } = useAdminPartners();
+  
   if (loading) return <Loader fixed />;
 
   return (
