+++ frontend/src/features/users/ForgotPassword.tsx
@@ -1,35 +1,18 @@
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
 import { Loader } from '@/components/Loader/Loader';
 import { Button } from '@/components/ui/button';
 import { UsersInput } from '@/features/users/components/UsersInput/UsersInput';
-import { selectForgotPasswordError, selectForgotPasswordLoading } from '@/features/users/usersSlice';
-import { forgotPassword } from '@/features/users/usersThunks';
 import { ArrowRightIcon } from '@heroicons/react/24/outline';
-import React, { type ChangeEvent, type FormEvent, useState } from 'react';
-import { toast } from 'sonner';
+import React from 'react';
+import {useForgotPassword} from '@/features/users/hooks/forgotPassword';
 
 export const ForgotPassword: React.FC = () => {
-  const dispatch = useAppDispatch();
-  const forgotError = useAppSelector(selectForgotPasswordError);
-  const forgotPasswordLoading = useAppSelector(selectForgotPasswordLoading);
-  const [email, setEmail] = useState('');
 
-  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const { value } = event.target;
-    setEmail(value);
-  };
-
-  const handleSubmit = async (event: FormEvent) => {
-    try {
-      event.preventDefault();
-
-      await dispatch(forgotPassword(email)).unwrap();
-      toast.success('Ссылка для сброса пароля отправлена на вашу почту.');
-    } catch (error) {
-      console.error(error);
-      toast.error(forgotError?.error || 'Что-то пошло не так.');
-    }
-  };
+  const {
+    forgotPasswordLoading,
+    email,
+    handleChange,
+    handleSubmit
+  } = useForgotPassword();
 
   return (
     <form onSubmit={handleSubmit}>
