+++ frontend/src/features/users/Register.tsx
@@ -1,148 +1,36 @@
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
-import { Loader } from '@/components/Loader/Loader';
-import { Button } from '@/components/ui/button';
-import { Checkbox } from '@/components/ui/checkbox';
-import { Label } from '@/components/ui/label';
-import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
-import { selectCategories, selectCategoriesFetching } from '@/features/category/categorySlice';
-import { fetchCategories } from '@/features/category/categoryThunks';
-import { UsersInput } from '@/features/users/components/UsersInput/UsersInput';
-import { selectRegisterError, selectRegisterLoading } from '@/features/users/usersSlice';
-import { register } from '@/features/users/usersThunks';
-import type { RegisterMutation } from '@/types/userTypes';
-import { ArrowLongRightIcon } from '@heroicons/react/24/outline';
-import React, { type ChangeEvent, type FormEvent, useEffect, useState } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
-import { toast } from 'sonner';
-
-const initialState: RegisterMutation = {
-  telephone: '',
-  password: '',
-  category: '',
-  fullName: '',
-  gender: '',
-  dateOfBirth: '',
-  email: '',
-};
+import {Loader} from '@/components/Loader/Loader';
+import {Button} from '@/components/ui/button';
+import {Checkbox} from '@/components/ui/checkbox';
+import {Label} from '@/components/ui/label';
+import {Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue} from '@/components/ui/select';
+import {UsersInput} from '@/features/users/components/UsersInput/UsersInput';
+import {ArrowLongRightIcon} from '@heroicons/react/24/outline';
+import React from 'react';
+import {Link} from 'react-router-dom';
+import {useRegister} from '@/features/users/hooks/register';
 
 export const Register: React.FC = () => {
-  const dispatch = useAppDispatch();
-  const loading = useAppSelector(selectRegisterLoading);
-  const categories = useAppSelector(selectCategories);
-  const categoriesFetching = useAppSelector(selectCategoriesFetching);
-  const error = useAppSelector(selectRegisterError);
-  const navigate = useNavigate();
-  const [registerMutation, setRegisterMutation] = useState(initialState);
-  const [confirmPassword, setConfirmPassword] = useState<string>('');
-  const [isRulesChecked, setIsRulesChecked] = useState({
-    rules: false,
-    personalData: false,
-  });
 
-  useEffect(() => {
-    if (error && error.errors) {
-      Object.values(error.errors).forEach((err) => {
-        toast.error(err.message);
-      });
-    }
-  }, [error]);
-
-  useEffect(() => {
-    dispatch(fetchCategories());
-  }, [dispatch]);
-
-  const handleDateChange = (event: ChangeEvent<HTMLInputElement>) => {
-    let value = event.target.value.replace(/\D/g, '');
-
-    if (value.length > 8) {
-      value = value.slice(0, 8);
-    }
-
-    let formattedDate = '';
-
-    if (value.length > 0) {
-      const day = value.slice(0, 2);
-      if (parseInt(day, 10) > 31) {
-        formattedDate += '31';
-      } else {
-        formattedDate += day;
-      }
-    }
-    if (value.length > 2) {
-      formattedDate += '.';
-      const month = value.slice(2, 4);
-      if (parseInt(month, 10) > 12) {
-        formattedDate += '12';
-      } else {
-        formattedDate += month;
-      }
-    }
-    if (value.length > 4) {
-      formattedDate += '.';
-      formattedDate += value.slice(4);
-    }
-
-    updateRegisterField('dateOfBirth', formattedDate);
-  };
-
-  const handleRulesChange = (value: boolean, id: string) => {
-    setIsRulesChecked((prev) => ({ ...prev, [id]: value }));
-  };
-
-  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const { id, value } = event.target;
-
-    if (id === 'telephone') {
-      const digitsOnly = value.replace(/\D/g, '');
-      let formattedPhone = digitsOnly;
-
-      if (digitsOnly.length > 1) {
-        formattedPhone = '0' + digitsOnly.slice(1, 4);
-      }
-      if (digitsOnly.length > 4) {
-        formattedPhone += ' ' + digitsOnly.slice(4, 7);
-      }
-      if (digitsOnly.length > 7) {
-        formattedPhone += ' ' + digitsOnly.slice(7, 10);
-      }
-
-      setRegisterMutation((prev) => ({ ...prev, telephone: formattedPhone }));
-      return;
-    }
-
-    updateRegisterField(id, value);
-  };
-
-  const handleSelectChange = (value: string, id: string) => {
-    const field = id === 'gender' ? 'gender' : 'category';
-    updateRegisterField(field, value);
-  };
-
-  const updateRegisterField = (field: string, value: string) => {
-    setRegisterMutation((prev) => ({ ...prev, [field]: value }));
-  };
-
-  const isFormValid = () => {
-    const isFilled =
-      Object.values(registerMutation).every((value) => value.trim() !== '') && confirmPassword.trim() !== '';
-    const passwordsMatch = registerMutation.password === confirmPassword;
-    const isRulesAccepted = Object.values(isRulesChecked).every((value) => value);
-
-    return isFilled && passwordsMatch && isRulesAccepted;
-  };
-
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    await dispatch(register(registerMutation)).unwrap();
-    setRegisterMutation(initialState);
-    navigate('/');
-  };
+  const {
+    loading,
+    categories,
+    categoriesFetching,
+    registerMutation,
+    confirmPassword,
+    setConfirmPassword,
+    handleDateChange,
+    handleSelectChange,
+    handleRulesChange,
+    handleChange,
+    isFormValid,
+    handleSubmit
+  } = useRegister();
 
   return (
     <form onSubmit={handleSubmit}>
       <section
         className='w-full my-14 py-10 mx-auto px-6 xs:max-w-[545px] xs:py-12 xs:px-10 rounded-3xl'
-        style={{ boxShadow: '0px 4px 100px 0px #00000017' }}
+        style={{boxShadow: '0px 4px 100px 0px #00000017'}}
       >
         <div className='mb-3'>
           <h1 className='font-bold text-[28px]'>Создать аккаунт.</h1>
@@ -212,7 +100,7 @@ export const Register: React.FC = () => {
             <Label htmlFor='gender'>Пол</Label>
             <Select value={registerMutation.gender} onValueChange={(value) => handleSelectChange(value, 'gender')}>
               <SelectTrigger className={'h-12 focus:ring-[#80BC41]'} id='gender'>
-                <SelectValue placeholder='Укажите ваш пол' />
+                <SelectValue placeholder='Укажите ваш пол'/>
               </SelectTrigger>
               <SelectContent>
                 <SelectGroup>
@@ -231,7 +119,7 @@ export const Register: React.FC = () => {
               onValueChange={(value) => handleSelectChange(value, 'category')}
             >
               <SelectTrigger className={'h-12 focus:ring-[#80BC41]'} id='category'>
-                <SelectValue placeholder='Выберите вашу категорию' />
+                <SelectValue placeholder='Выберите вашу категорию'/>
               </SelectTrigger>
               <SelectContent>
                 <SelectGroup>
@@ -276,7 +164,7 @@ export const Register: React.FC = () => {
           disabled={!isFormValid()}
         >
           Зарегистрироваться
-          {loading ? <Loader /> : <ArrowLongRightIcon style={{ width: 40, height: 40 }} strokeWidth={1} />}
+          {loading ? <Loader/> : <ArrowLongRightIcon style={{width: 40, height: 40}} strokeWidth={1}/>}
         </Button>
 
         <Link to='/login' className='text-sm block text-center text-black/50 w-fit mx-auto'>
