+++ frontend/src/features/users/ResetPassword.tsx
@@ -1,44 +1,19 @@
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
 import { Loader } from '@/components/Loader/Loader';
 import { Button } from '@/components/ui/button';
 import { UsersInput } from '@/features/users/components/UsersInput/UsersInput';
-import { selectResetPasswordError, selectResetPasswordLoading } from '@/features/users/usersSlice';
-import { resetPassword } from '@/features/users/usersThunks';
 import { ArrowRightIcon } from '@heroicons/react/24/outline';
-import React, { type ChangeEvent, type FormEvent, useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
-import { toast } from 'sonner';
+import React from 'react';
+import {useResetPassword} from '@/features/users/hooks/resetPassword';
 
 export const ResetPassword: React.FC = () => {
-  const navigate = useNavigate();
-  const dispatch = useAppDispatch();
-  const { token } = useParams() as { token: string };
-  const resetPasswordLoading = useAppSelector(selectResetPasswordLoading);
-  const resetPasswordError = useAppSelector(selectResetPasswordError);
-  const [passwords, setPasswords] = useState({
-    password: '',
-    confirmPassword: '',
-  });
 
-  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const { id, value } = event.target;
-    setPasswords({ ...passwords, [id]: value });
-  };
-
-  const passwordMatch = passwords.password !== passwords.confirmPassword;
-
-  const handleSubmit = async (event: FormEvent) => {
-    try {
-      event.preventDefault();
-
-      await dispatch(resetPassword({ password: passwords.password, token })).unwrap();
-      toast.success('Пароль успешно изменен.');
-      navigate('/login');
-    } catch (error) {
-      console.error(error);
-      toast.error(resetPasswordError?.error || 'Что-то пошло не так.');
-    }
-  };
+  const {
+    resetPasswordLoading,
+    passwords,
+    passwordMatch,
+    handleChange,
+    handleSubmit 
+  } = useResetPassword();
 
   return (
     <form onSubmit={handleSubmit}>
